# perf: optimize factory performance