--- a/factory/exported_gen.go
+++ b/factory/exported_gen.go
@@ -16,16 +16,18 @@ var std = New()
 // StandardFactory returns standard factory.
 func StandardFactory() *Factory { return std }
 
-// CreateMesg creates new message based on defined messages in the factory. If not found, it returns new message with "unknown" name.
+// CreateMesg creates new message based on defined messages in the factory. If not found, it returns new unknown message.
 //
-// This will create a shallow copy of the Fields, so changing any value declared in Field's FieldBase is prohibited (except fot the unknown field).
-// If you want a deep copy of the mesg, create it by calling mesg.Clone().
+// This will create a shallow copy of the Fields, so changing any value declared in Field's FieldBase is prohibited (except in case of unknown field).
+// If you want a deep copy of the mesg, clone it by calling mesg.Clone().
+//
+// NOTE: This method is not used by either the decoder or the encoder, and the data will only be populated once upon the first invocation.
 func CreateMesg(num typedef.MesgNum) proto.Message {
 	return std.CreateMesg(num)
 }
 
-// CreateMesgOnly is similar to CreateMesg, but it sets Fields to nil. This is useful when we plan to fill these values ourselves
-// to avoid unnecessary malloc when cloning them, as they will be removed anyway. For example, the decoding process will populate them with decoded data.
+// CreateMesgOnly creates new message without predefined fields. This is useful when we plan to fill these values ourselves
+// e.g the decoding process will populate them with decoded data.
 func CreateMesgOnly(num typedef.MesgNum) proto.Message {
 	return std.CreateMesgOnly(num)
 }