- Use [256]proto.Field in the predefined message in the factory for fast look up, this ensure O(1) look up when creating Field instead of looping to get the Field like previous version.
- Since the predefined messages in factory is now using fixed array, the behavior of `CreateMesg` is slightly changed since we need to populate the slice for the returned message's fields. It will populate the data once during the first invocation, since this method is not used either by the decoder or the encoder it should be fine, method documentation is updated to make users aware about this change in case users want to use this method.

```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/decoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
         │   old.txt   │               new.txt               │
         │   sec/op    │   sec/op     vs base                │
Decode-4   126.4m ± 1%   108.6m ± 1%  -14.08% (p=0.000 n=10)

         │   old.txt    │            new.txt             │
         │     B/op     │     B/op      vs base          │
Decode-4   67.39Mi ± 0%   67.39Mi ± 0%  ~ (p=0.361 n=10)

         │   old.txt   │             new.txt             │
         │  allocs/op  │  allocs/op   vs base            │
Decode-4   900.0k ± 0%   900.0k ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

```